Move Image, Layout, Links, Lists components into directories
