+++ assets/javascripts/kitten/components/images/image-with-caption/index.js
