+++ assets/javascripts/kitten/components/layout/marger/test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
-import { Marger } from '../../components/layout/marger'
+import { Marger } from '../../../components/layout/marger'
 
 describe('<Marger />', () => {
   describe('by default', () => {
