+++ assets/javascripts/kitten/components/layout/separator/index.js
