+++ assets/javascripts/kitten/components/layout/vertical-stroke/index.js
