+++ assets/javascripts/kitten/components/links/external-rich-link/index.js
