+++ assets/javascripts/kitten/components/lists/nav-list/index.js
